dont add internal domain as self hosted

+++ images/03.serviceImages/dns/home/add_service.sh
@@ -10,11 +10,14 @@ load_service_hash_to_environment
 
 	if ! test -z ${domain_name}
 	 then
+	  if ! test ${domain_name} = engines.internal
+	   then
 	 	cat  /etc/bind/templates/config_file_zone_entry.tmpl | sed " /DOMAIN/s//${domain_name}/g" > /home/bind/engines/domains/${domain_name}
 	 	cat /etc/bind/templates/selfhosted.tmpl | sed "/DOMAIN/s//${domain_name}/g" | sed "/IP/s//${ip}/g" > /home/bind/engines/zones/named.conf.${domain_name}
 	 	cat /home/bind/engines/domains/* > /home/bind/engines/domains.hosted
 	 	kill -HUP `cat /var/run/named/named.pid`
 	 	exit
+	   fi
 	 fi
 
 #FIXME make engines.internal settable
